MDL-51210 CACHING Explicitly cast int for Memcache server port
Fix notice of: "A non well formed numeric value encountered" when multiple
memcache servers are configured for a single cache store in
Moodle Unified Caching (MUC).

Fix casts Memcache server port explicitly as an integer when port value
is parsed from the Memcache connection string. Previously port
number was cast as a string as a result of the substr() extraction.

+++ lib/classes/session/util.php
@@ -63,7 +63,7 @@ abstract class util {
             $pos  = strrpos($part, ':');
             if ($pos !== false) {
                 $host = substr($part, 0, $pos);
-                $port = substr($part, ($pos + 1));
+                $port = (int) substr($part, ($pos + 1));
             } else {
                 $host = $part;
                 $port = 11211;
